chore: move app-wrapper to components and fix ESLint rule violations
