+++ i18n/en.pot
@@ -5,8 +5,20 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-08-08T12:23:45.715Z\n"
-"PO-Revision-Date: 2025-08-08T12:23:45.716Z\n"
+"POT-Creation-Date: 2025-08-11T13:30:16.690Z\n"
+"PO-Revision-Date: 2025-08-11T13:30:16.691Z\n"
+
+msgid "ERROR"
+msgstr "ERROR"
+
+msgid "Loading..."
+msgstr "Loading..."
+
+msgid "Hello {{name}}"
+msgstr "Hello {{name}}"
+
+msgid "Welcome to DHIS2 with TypeScript!"
+msgstr "Welcome to DHIS2 with TypeScript!"
 
 msgid "User organisation unit"
 msgstr "User organisation unit"
@@ -115,18 +127,6 @@ msgstr "This year"
 
 msgid "Last year"
 msgstr "Last year"
-
-msgid "ERROR"
-msgstr "ERROR"
-
-msgid "Loading..."
-msgstr "Loading..."
-
-msgid "Hello {{name}}"
-msgstr "Hello {{name}}"
-
-msgid "Welcome to DHIS2 with TypeScript!"
-msgstr "Welcome to DHIS2 with TypeScript!"
 
 msgid ""
 "Track or compare changes over time. Recommend period as category. (adjust "
