+++ src/app.tsx
@@ -2,8 +2,8 @@ import i18n from '@dhis2/d2-i18n'
 import { CssVariables } from '@dhis2/ui'
 import cx from 'classnames'
 import React, { FC } from 'react'
-import { AppWrapper } from './app-wrapper'
 import classes from './app.module.css'
+import { AppWrapper } from './components/app-wrapper'
 import { useRtkQuery, useSystemSettings } from './hooks'
 import { Toolbar } from '@components/toolbar/toolbar'
 import type { MeDto } from '@types'
