+++ src/types/rtk-query-hooks.ts
@@ -2,8 +2,7 @@ import type {
     QueryActionCreatorResult,
     QueryDefinition,
 } from '@reduxjs/toolkit/query'
-import type { CustomBaseQueryFn, EngineArgs } from '../api/custom-base-query'
-import type { EngineError } from '../api/parse-engine-error'
+import type { CustomBaseQueryFn, EngineArgs, EngineError } from '@api'
 
 // Probably we don't want this manual type
 export type UseRtkQueryResult<T> = {
